pull out types
